pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cli-weather-dashboard"
version = "1.0.0"
description = "A CLI tool to fetch and display weather data"
authors = [{name = "Antigravity", email = "antigravity@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "rich==13.7.0"
]

[project.scripts]
weather = "weather_cli.weather_dashboard:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["weather_cli"]